requirements.txt
langchain-groq
langgraph
pydantic
python-dotenv
pymupdf
tabulate